Cleaned the folder
