+++ client/src/components/Homepage/SignUp.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const SignUp = () => {
+  return (
+    <div>SignUp</div>
+  )
+}
+
+export default SignUp
