--- client/src/components/SignUp.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const SignUp = () => {
-  return (
-    <div>SignUp</div>
-  )
-}
-
-export default SignUp
